pyproject.toml
[tool.poetry]
name = "todo_rest"
version = "0.1.0"
description = ""
authors = ["ZZaRRyaDD <odddellarobea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.8.0"
django-cors-headers = "^3.10.0"
python-decouple = "^3.5"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
